Renamed the API caller file to match what it was supposed to be called in the exercise instructions!
